+++ src/Complexity/Complexity.php
@@ -9,6 +9,8 @@
  */
 namespace SebastianBergmann\Complexity;
 
+use function str_contains;
+
 /**
  * @psalm-immutable
  */
@@ -48,5 +50,15 @@ final class Complexity
     public function cyclomaticComplexity(): int
     {
         return $this->cyclomaticComplexity;
+    }
+
+    public function isFunction(): bool
+    {
+        return !$this->isMethod();
+    }
+
+    public function isMethod(): bool
+    {
+        return str_contains($this->name, '::');
     }
 }